+++ src/main/java/model/card/SuitType.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 public enum SuitType {
