+++ src/main/java/model/participant/Players.java
@@ -1,4 +1,4 @@
-package model;
+package model.participant;
 
 import java.util.List;
