+++ src/test/java/model/GameManagerTest.java
@@ -1,9 +1,12 @@
 package model;
 
 import java.util.List;
+
+import model.participant.Dealer;
+import model.participant.Player;
+import model.participant.Players;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
