+++ src/test/java/model/card/CardTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.card;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
