+++ src/test/java/model/participant/DealerTest.java
@@ -1,8 +1,13 @@
-package model;
+package model.participant;
 
 import java.util.List;
 import java.util.Map;
 
+import model.card.Card;
+import model.card.CardDeck;
+import model.card.RankType;
+import model.card.SuitType;
+import model.score.MatchType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
