+++ src/test/java/model/participant/NicknameTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.participant;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
