+++ src/test/java/model/participant/PlayersTest.java
@@ -1,10 +1,13 @@
-package model;
+package model.participant;
 
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
 import java.util.stream.Stream;
 
+import model.card.Card;
+import model.card.RankType;
+import model.card.SuitType;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
