+++ src/test/java/model/score/ScoreTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.score;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
